+++ controllers/users.js
@@ -1,9 +1,9 @@
 const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
 const User = require('../models/user');
-const { JWT_DEV } = require('../config');
+const { JWT_SECRET } = require('../config');
+// const { NODE_ENV, JWT_SECRET } = process.env;
 
-const { NODE_ENV, JWT_SECRET } = process.env;
 const {
   BadRequestError,
   UnauthorizedError,
@@ -52,7 +52,7 @@ module.exports.login = (req, res, next) => {
 
   return User.findUserByCredentials(email, password)
     .then((user) => {
-      const token = jwt.sign({ _id: user._id }, NODE_ENV === 'production' ? JWT_SECRET : JWT_DEV, { expiresIn: '7d' });
+      const token = jwt.sign({ _id: user._id }, JWT_SECRET, { expiresIn: '7d' });
 
       res
         .cookie('jwt', token, {
