+++ index.js
@@ -3,29 +3,23 @@ const mongoose = require('mongoose');
 const bodyParser = require('body-parser');
 const cookieParser = require('cookie-parser');
 const helmet = require('helmet');
-const rateLimit = require('express-rate-limit');
-require('dotenv').config();
+// require('dotenv').config();
 
 const { errors } = require('celebrate');
 const { PORT, DATABASE_URL } = require('./config');
 const routes = require('./routes');
-
+const limiter = require('./middlewares/rate-limiter');
 const { requestLogger, errorLogger } = require('./middlewares/logger');
+const errorHandler = require('./middlewares/error-500-handler');
 
 const app = express();
 app.listen(PORT);
-
-const limiter = rateLimit({
-  windowMs: 15 * 60 * 1000, // 15 minutes
-  max: 100, // limit each IP to 100 requests per windowMs
-});
 
 app.use(limiter);
 app.use(helmet());
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
 
-// DATABASE_URL??
 mongoose.connect(DATABASE_URL, {
   useNewUrlParser: true,
   useCreateIndex: true,
@@ -36,22 +30,9 @@ mongoose.connect(DATABASE_URL, {
 app.use(requestLogger);
 app.use(cookieParser());
 app.use(routes);
-app.use(errorLogger);
-app.use(errors()); // обработчик ошибок celebrate
-
-// eslint-disable-next-line no-unused-vars
-app.use((err, req, res, next) => {
-  const { statusCode = 500, message } = err;
-
-  res
-    .status(statusCode)
-    .send({
-      message: statusCode === 500
-        ? 'На сервере произошла ошибка'
-        : message,
-    });
-});
-
 app.all('*', (req, res) => {
   res.status(404).send({ message: 'Запрашиваемый ресурс не найден' });
 });
+app.use(errorLogger);
+app.use(errors()); // обработчик ошибок celebrate
+app.use(errorHandler);
