+++ middlewares/error-500-handler.js
@@ -1,0 +1,14 @@
+const errorHandler = (err, req, res, next) => {
+  const { statusCode = 500, message } = err;
+
+  res
+    .status(statusCode)
+    .send({
+      message: statusCode === 500
+        ? 'На сервере произошла ошибка'
+        : message,
+    });
+  next();
+};
+
+module.exports = errorHandler;
